modifications according to dschult's suggestions
